requirements.txt
langgraph==0.3.34
langchain-openai==0.3.16
openai==1.77.0
beautifulsoup4==4.13.3
requests==2.32.3
python-docx==1.0.0
python-dotenv==1.0.1
selenium==4.15.0
webdriver-manager==4.0.1
aiohttp==3.11.11
nltk==3.9.1
spacy==3.7.2
pandas==2.2.3
numpy==1.26.4
pydantic==2.10.6
typing-extensions==4.12.2
matplotlib>=3.5.0

# Add Supabase client (compatible version)
supabase>=2.18.0
